Move AccountAutoDiscoveryStateMapper to autodiscovery package
